Split clients file in client directory and expect the developer to require its client himself instead of loading them all.
