+++ Gemfile
@@ -1,7 +1,5 @@
 source 'http://rubygems.org'
 
-gem 'oauth'
-gem 'simple_oauth'
 gem 'faraday'
 gem 'builder'
