+++ Rakefile
@@ -8,9 +8,7 @@ begin
     gem.summary = %Q{The most complete Ruby wrapper for youtube api's}
     gem.description = %Q{Upload, delete, update, comment on youtube videos all from one gem.}
     gem.email = "kylejginavan@gmail.com"
-    gem.homepage = "http://github.com/kylejginavan/youtube_it"
-    gem.add_dependency('oauth','>=0.4.4')
-    gem.add_dependency('simple_oauth', '>=0.1.5')    
+    gem.homepage = "http://github.com/kylejginavan/youtube_it" 
     gem.add_dependency('faraday','>=0.7.3')    
     gem.add_dependency('builder')
     gem.authors = ["kylejginavan","chebyte", "mseppae"]
